pyproject.toml
[project]
name = "data-warehouse-apscheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.15",
    "apscheduler>=3.11.0",
    "ciso8601>=2.3.3",
    "colorama>=0.4.6",
    "dotenv>=0.9.9",
    "fastapi>=0.115.0",
    "krex>=0.17.0",
    "motor>=3.7.1",
    "msgspec>=0.19.0",
    "pandas>=2.3.2",
    "psutil>=5.9.8",
    "psycopg[binary]>=3.2",
    "pymongo>=4.14.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.32.0",
    "websockets>=15.0.1",
]
